refactor: implement JSON-Schema normalization for strict structured output
