+++ tests/test_openai.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel
 
 from lmdk.datatypes import CompletionRequest, UserMessage
 from lmdk.provider import RawResponse
-from lmdk.providers.openai import OPENAI_API_URL, OpenAIProvider, OpenaiProvider, _prepare_schema
+from lmdk.providers.openai import OPENAI_API_URL, OpenAIProvider, OpenaiProvider
 
 # ---------------------------------------------------------------------------
 # Helpers
@@ -178,49 +178,6 @@ class TestBuildPayload:
         request = _make_request(output_schema=Person)
         payload = OpenaiProvider._build_payload(request, stream=True)
         assert "text" not in payload
-
-
-# ---------------------------------------------------------------------------
-# _prepare_schema
-# ---------------------------------------------------------------------------
-
-
-class TestPrepareSchema:
-    def test_adds_openai_strict_object_constraints(self):
-        result = _prepare_schema(Person.model_json_schema())
-        assert result["additionalProperties"] is False
-        assert result["required"] == ["name", "age"]
-
-    def test_handles_nested_defs_and_defaults(self):
-        result = _prepare_schema(Recipe.model_json_schema())
-        ingredient = result["$defs"]["Ingredient"]
-        assert result["additionalProperties"] is False
-        assert ingredient["additionalProperties"] is False
-        assert ingredient["required"] == ["name", "quantity", "unit"]
-        assert "default" not in ingredient["properties"]["unit"]
-
-    def test_does_not_mutate_original_schema(self):
-        schema = Recipe.model_json_schema()
-        _prepare_schema(schema)
-        assert "additionalProperties" not in schema
-        assert "default" in schema["$defs"]["Ingredient"]["properties"]["unit"]
-
-    def test_handles_arrays_and_combinators(self):
-        schema = {
-            "type": "object",
-            "properties": {
-                "items": {
-                    "type": "array",
-                    "items": {
-                        "anyOf": [{"type": "object", "properties": {"name": {"type": "string"}}}]
-                    },
-                }
-            },
-        }
-        result = _prepare_schema(schema)
-        nested = result["properties"]["items"]["items"]["anyOf"][0]
-        assert nested["additionalProperties"] is False
-        assert nested["required"] == ["name"]
 
 
 # ---------------------------------------------------------------------------
